+++ src/index.js
@@ -5,7 +5,7 @@ import {CONTENT_DIRECTORY, BLOG_POSTS_DIRECTORY, OUTPUT_DIRECTORY} from './confi
 import {loadManyBlogPostSourceFiles} from './blog-post/load-blog-post-source-file.js';
 import {loadManyBlogPosts} from './blog-post/load-blog-post.js';
 import {sortByDateCreatedDescending} from './blog-post/sort-blog-post.js';
-import {groupBlogPostsByTag, groupBlogPostsByYearAndMonth} from './blog-post/group-blog-posts.js';
+import {groupArticlesByTag, groupArticlesByYearAndMonth} from './_shared/group-articles.js';
 import {loadTidbits} from './load-tidbit/load-tidbit.js';
 import {loadManyTidbitSourceFiles} from './load-tidbit/load-tidbit-source-file.js';
 import {renderTemplate} from './_deps/render-template.js';
@@ -52,11 +52,11 @@ const generatePost = async (post) => {
   await fs.promises.writeFile(destFilename, renderedFile);
 }
 
-import {renderAndWriteTidbitPages, renderAndWriteTidbitsIndexPage} from './render-tidbit/render-page.js';
-const generateTidbitsPages = async (tidbits) => {
-  await renderAndWriteTidbitsIndexPage()(tidbits, defaultRenderParams);
-  await renderAndWriteTidbitPages()(tidbits, defaultRenderParams);
-};
+import {renderAndWriteTidbitPages, renderAndWriteTidbitsIndexPage, renderAndWriteTidbitTagPages} from './render-tidbit/render-page.js';
+const generateSingleTidbitsPages = (tidbits) => renderAndWriteTidbitPages()(tidbits, defaultRenderParams);
+const generateTidbitsIndexPage = (tidbits) => renderAndWriteTidbitsIndexPage()(tidbits, defaultRenderParams);
+const generateTidbitTagPages = (tidbitGroups) => renderAndWriteTidbitTagPages()(tidbitGroups, defaultRenderParams);
+
 
 const aboutIndexPage = async () => {
   const destDir = path.join(OUTPUT_DIRECTORY, 'about');
@@ -115,7 +115,7 @@ const generateMonthPage = async (group) => {
   const destDir = path.join(OUTPUT_DIRECTORY, 'blog', yearAndMonth.replace('-', '/'));
   await fs.promises.mkdir(destDir, {recursive: true});
   const destFilename = path.join(destDir, 'index.html');
-  const renderedFile = renderTemplate('blog/month.html', {...defaultRenderParams, yearAndMonth, posts: group.blogPosts});
+  const renderedFile = renderTemplate('blog/month.html', {...defaultRenderParams, yearAndMonth, articles: group.articles});
   await fs.promises.writeFile(destFilename, renderedFile);
 };
 
@@ -153,14 +153,15 @@ const loadPosts = async sourceFiles => {
   console.time('Load blog posts');
   const posts = await loadPosts(sourceFiles);
   console.timeEnd('Load blog posts');
+
   console.time('Relate and group posts');
   posts.forEach(post => post.relatedPosts = findRelatedPosts(post, posts));
-  const blogPostsGroupedByTag = groupBlogPostsByTag(posts);
+  const blogPostsGroupedByTag = groupArticlesByTag(posts, '/blog/tag');
   const sortAlphabeticallyByTag = (group1, group2) => group1.tagSlug > group2.tagSlug ? 1 : -1;
   const groupedBlogPosts = {
     byTag: blogPostsGroupedByTag,
     byTagSortedAlphabetically: [...blogPostsGroupedByTag].sort(sortAlphabeticallyByTag),
-    byMonth: groupBlogPostsByYearAndMonth(posts),
+    byMonth: groupArticlesByYearAndMonth(posts, '/blog'),
   };
   defaultRenderParams.groupedBlogPosts = groupedBlogPosts;
   console.timeEnd('Relate and group posts');
@@ -171,21 +172,37 @@ const loadPosts = async sourceFiles => {
   const tidbits = await loadTidbits()(tidbitSourceFiles);
   console.timeEnd('Load tidbits');
 
+  console.time('Relate and group tidbits');
+  const tidbitsGroupedByTag = groupArticlesByTag(tidbits, '/tidbits/tag');
+  const groupedTidbits = {
+    byTag: tidbitsGroupedByTag,
+    byTagSortedAlphabetically: [...tidbitsGroupedByTag].sort(sortAlphabeticallyByTag),
+    byMonth: groupArticlesByYearAndMonth(tidbits, '/tidbits'),
+  };
+  defaultRenderParams.groupedTidbits = groupedTidbits;
+  console.timeEnd('Relate and group tidbits');
+
   console.log('\nBuilding pages\n========');
   await runAndTimeIt('Home page', () => generateHomePage(posts.excludingDrafts(), tidbits));
   // blog
-  await runAndTimeIt(`All blog posts (${posts.length})`, () => Promise.all(posts.map(generatePost)));
-  await runAndTimeIt('Blog overview page', () => generateBlogOverviewPage(posts.excludingDrafts()));
+  console.log('Blog');
+  await runAndTimeIt(`  all posts (${posts.length})`, () => Promise.all(posts.map(generatePost)));
+  await runAndTimeIt('  /blog page', () => generateBlogOverviewPage(posts.excludingDrafts()));
+  await runAndTimeIt(`  tags pages (${groupedBlogPosts.byTag.length})`, () => generateTagPages(groupedBlogPosts.byTag));
+  await runAndTimeIt(`  month pages (${groupedBlogPosts.byMonth.length})`, () => generateMonthPages(groupedBlogPosts.byMonth));
 
-  await runAndTimeIt(`All tags pages (${groupedBlogPosts.byTag.length})`, () => generateTagPages(groupedBlogPosts.byTag));
-  await runAndTimeIt(`All month pages (${groupedBlogPosts.byMonth.length})`, () => generateMonthPages(groupedBlogPosts.byMonth));
+  console.log('Tidbit');
+  await runAndTimeIt(`  /tidbits page`, () => generateTidbitsIndexPage(tidbits));
+  await runAndTimeIt(`  all pages (${tidbits.length})`, () => generateSingleTidbitsPages(tidbits));
+  await runAndTimeIt(`  tags pages (${groupedTidbits.byTag.length})`, () => generateTidbitTagPages(groupedTidbits.byTag));
   await runAndTimeIt('About pages', () => generateAboutPages());
-  await runAndTimeIt(`Tidbit pages (${tidbits.length})`, () => generateTidbitsPages(tidbits));
   await runAndTimeIt('Projects page', () => generateProjectsPage());
   await runAndTimeIt('Projects plan page', () => generateProjectsPlanPage());
+
+  console.log('HTTP pages');
   const oldUrlsCount = posts.reduce((prev, cur) => cur.oldUrls.length + prev, 0);
-  await runAndTimeIt(`All 301 pages (${oldUrlsCount})`, () => Promise.all(posts.map(generate301Pages)));
-  await runAndTimeIt('404 page', () => generate404Page(posts.slice(0, 5)));
+  await runAndTimeIt(`  301 pages (${oldUrlsCount})`, () => Promise.all(posts.map(generate301Pages)));
+  await runAndTimeIt('  404 page', () => generate404Page(posts.slice(0, 5)));
   
   console.log('-----');
   console.timeEnd('Overall');