+++ src/render-blog/render-page.slow-spec.js
@@ -3,10 +3,6 @@ import {assertThat, containsString, hasItem, matchesPattern} from 'hamjest';
 import {BlogPost} from "../blog-post/BlogPost.js";
 import {renderAndWriteTagPages} from "./render-page.js";
 
-/**
- * @typedef {import('../blog-post/types').BlogPostsByTag} BlogPostsByTag
- */
-
 // TODO THIS is really ugly, that we have to inject that every time.
 // Maybe intro a `DefaultRenderParameters.empty()` or something.
 const renderParams = {navigationItems: [], groupedBlogPosts: {byTag: [], byMonth: []}};
@@ -32,14 +28,22 @@ const createBlogPost = (overrideData = {}) => {
   return BlogPost.withRawData({...defaultRawBlogPostData, ...overrideData});
 }
 
+/**
+ * @param props {{tagSlug: Slug, articles: Article[], gradientWidthInPercent?: number, url?: string}}
+ * @return {ArticlesGroupedByTag}
+ */
+const newGroup = ({tagSlug, articles, gradientWidthInPercent = 42, url = '/irrelevant'}) => {
+  return {tagSlug, articles, gradientWidthInPercent, url};
+};
+
 describe('Render blog pages', () => {
   describe('GIVEN some blog posts WHEN rendering them', () => {
     describe('THEN render a page per tag', () => {
       /**
-       * @param groups {BlogPostsByTag[]}
+       * @param groups {ArticlesGroupedByTag[]}
        * @return {Promise<string>}
        */
-      const renderTagPage = async (groups = [{tagSlug: 'one', blogPosts: [createBlogPost()], gradientWidthInPercent: 100}]) => {
+      const renderTagPage = async (groups = [newGroup({tagSlug: 'one', articles: [createBlogPost()]})]) => {
         /** @type string */
         let writtenToFile = '';
         /**
@@ -54,7 +58,7 @@ describe('Render blog pages', () => {
 
       it('AND write "one" tag`s page to "/blog/tag/one/index.html"', async () => {
         const groups = [
-          {tagSlug: 'one', blogPosts: [createBlogPost()], gradientWidthInPercent: 100}
+          newGroup({tagSlug: 'one', articles: [createBlogPost()]}),
         ];
         /** @type Filename[] */
         const writtenToFilenames = [];
@@ -75,14 +79,14 @@ describe('Render blog pages', () => {
       });
       it('AND render the post`s headlines as H2', async () => {
         const groups = [
-          {tagSlug: 'one', blogPosts: [createBlogPost({headlineAsHtml: 'One tagged Blog-Post'})], gradientWidthInPercent: 100}
+          newGroup({tagSlug: 'one', articles: [createBlogPost({headlineAsHtml: 'One tagged Blog-Post'})]}),
         ];
         const writtenToFile = await renderTagPage(groups);
         assertThat(writtenToFile, matchesPattern(/<h2.*>.*One tagged Blog-Post.*<\/h2>/gms));
       });
       it('AND renders "tagged with: #one #two" under the post', async () => {
         const groups = [
-          {tagSlug: 'one', blogPosts: [createBlogPost({tags: ['one', 'two']})], gradientWidthInPercent: 100},
+          newGroup({tagSlug: 'one', articles: [createBlogPost({tags: ['one', 'two']})]}),
         ];
         const writtenToFile = await renderTagPage(groups);
         assertThat(writtenToFile, containsString('tagged with: #one #two'));
@@ -90,4 +94,3 @@ describe('Render blog pages', () => {
     });
   });
 });
-