+++ src/render-tidbit/render-page.js
@@ -32,3 +32,29 @@ export const renderAndWriteTidbitPages = ({writeFile = writeOutputFile, renderPa
   const pageWriters = tidbits.map(t => writeFile(t.url + 'index.html', renderPage({...renderParams, tidbit: t})));
   await Promise.all(pageWriters);
 };
+
+/**
+ * @param data {PlainObject}
+ * @return {string}
+ */
+const renderTagPage = (data) => renderTemplate('tidbit/tag.html', data);
+
+/**
+ * @param deps? {TagPageProductionDependencies}
+ * @return {function(ArticlesGroupedByTag[], PlainObject): Promise<void>}
+ */
+export const renderAndWriteTidbitTagPages = ({writeFile = writeOutputFile, renderPage = renderTagPage} = {}) => async (groups, renderParams) => {
+  /**
+   * @param tagSlug {string}
+   * @return {string}
+   */
+  const destFilename = tagSlug => `/tidbits/tag/${tagSlug}/index.html`;
+  /**
+   * @param group {ArticlesGroupedByTag}
+   * @return {Promise<void>}
+   */
+  const writeGroup = group => writeFile(destFilename(group.tagSlug), renderPage({...renderParams, tag: group.tagSlug, articles: group.articles}));
+  const pageWriters = groups.map(writeGroup);
+  await Promise.all(pageWriters);
+}
+