+++ src/render-tidbit/render-page.slow-spec.js
@@ -5,7 +5,11 @@ import {renderAndWriteTidbitPages, renderAndWriteTidbitsIndexPage} from './rende
 
 // TODO THIS is really ugly, that we have to inject that every time.
 // Maybe intro a `DefaultRenderParameters.empty()` or something.
-const renderParams = {navigationItems: [], groupedBlogPosts: {byTag: [], byMonth: []}};
+const renderParams = {
+  navigationItems: [],
+  groupedBlogPosts: {byTag: [], byMonth: []},
+  groupedTidbits: {byTag: [], byMonth: []},
+};
 
 const defaultRawTidbitData = {
   abstract: '',